# Conflicts:
#	App/Controllers/CategoriesController.php
#	App/Controllers/CommentsController.php
#	App/Controllers/ProductsController.php
#	App/Controllers/ReportsController.php
#	App/Controllers/SessionsController.php
#	App/Controllers/UsersController.php
#	App/System/App.php
#	App/System/Auth.php
#	App/Views/pages/admin/users_add.twig
#	App/Views/pages/admin/users_edit.twig
#	App/Views/pages/categories_add.twig
#	App/Views/pages/categories_edit.twig
#	App/Views/pages/dashboard.twig
#	App/Views/pages/products_add.twig
#	App/Views/pages/products_edit.twig
#	App/Views/pages/registration.twig
#	App/Views/pages/reports_add.twig
#	App/Views/pages/signin.twig

+++ App/Controllers/CategoriesController.php
@@ -3,6 +3,7 @@ namespace App\Controllers;
 
 use App\Models\ProductsModel;
 use \App\System\App;
+use \App\System\Auth;
 use \App\System\FormValidator;
 use \App\System\Settings;
 use \App\Controllers\Controller;
@@ -25,7 +26,7 @@ class CategoriesController extends Controller {
     }
 
     public function add() {
-        if(!empty($_POST)) {
+        if(!empty($_POST) && Auth::checkCSRF($_POST["token"])) {
             $title       = isset($_POST['title']) ? $_POST['title'] : '';
             $description = isset($_POST['description']) ? $_POST['description'] : '';
 
@@ -57,6 +58,7 @@ class CategoriesController extends Controller {
                     ]
                 ]);
             }
+            App::redirect('categories');
         }
 
         else {
@@ -69,7 +71,7 @@ class CategoriesController extends Controller {
     }
 
     public function edit($id) {
-        if(!empty($_POST)) {
+        if(!empty($_POST) && Auth::checkCSRF($_POST["token"])) {
             $title       = isset($_POST['title']) ? $_POST['title'] : '';
             $description = isset($_POST['description']) ? $_POST['description'] : '';
 
@@ -133,7 +135,7 @@ class CategoriesController extends Controller {
         $model2 = new ProductsModel($_COOKIE['user']);
         $products = $model2->getProductsByCategoryId($id);
         
-        if(!empty($_POST)) {
+        if(!empty($_POST) && Auth::checkCSRF($_POST["token"])) {
             foreach($products as $product){
                 $model2->delete($product->id);
             }
